[vs]VOQ Switch objects initialiation
Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>

The changes include association of local port id to sytems port that corresponds
to this local port oid. In real SAI, mapping between system port and local port is
done in run-time on-demand (when GET is done) by matching the system ports's
<switch_id, core index, core port index> tuple. In VOQ switches, each port blongs
to a switch core and is assigned an index. This mapping is done in hardware configurations.
For VS, the mapping is emulated via a new file coreportindexmap.ini. This file is
made available in /usr/share/sonic/hwsku in the same way how lanemap.ini is made avaialble.
Loading and parsing of coreportindexmap.ini is done in similar way as it is done for
lanemap.ini.  While emulating configuration of system ports, if a system port is found to
be a local port (determined using its switch_id), the local port oid corresponding
to the system port is retrieved from m_port_list and local port atribute is set in the
sytem port object. This is used by orchagent (portsorch) while adding system ports
information to portsorch local data structures.

Refer https://github.com/Azure/sonic-swss/pull/1431